requirements.txt
mistralai
python-dotenv
requests
pybase64
pillow
streamlit